+++ unix_integration/nss_sparkle_common/src/core.rs
@@ -1,3 +1,6 @@
+use libnss::group::Group;
+use libnss::interop::Response;
+use libnss::passwd::Passwd;
 use sparkle_unix_common::client_sync::DaemonClientBlocking;
 use sparkle_unix_common::constants::{SYSTEM_GROUP_PATH, SYSTEM_PASSWD_PATH};
 use sparkle_unix_common::unix_config::PamNssConfig;
@@ -5,9 +8,6 @@ use sparkle_unix_common::unix_passwd::{
     read_etc_group_file, read_etc_passwd_file, EtcGroup, EtcUser,
 };
 use sparkle_unix_common::unix_proto::{ClientRequest, ClientResponse, NssGroup, NssUser};
-use libnss::group::Group;
-use libnss::interop::Response;
-use libnss::passwd::Passwd;
 use std::ops::{Deref, DerefMut};
 use std::sync::RwLock;
