+++ unix_integration/nss_sparkle_common/src/hooks.rs
@@ -1,8 +1,8 @@
 use crate::core::{self, RequestOptions};
-use sparkle_unix_common::constants::DEFAULT_CONFIG_PATH;
 use libnss::group::{Group, GroupHooks};
 use libnss::interop::Response;
 use libnss::passwd::{Passwd, PasswdHooks};
+use sparkle_unix_common::constants::DEFAULT_CONFIG_PATH;
 
 pub struct SparklePasswd;
