+++ unix_integration/nss_sparkle_common/src/tests.rs
@@ -1,6 +1,6 @@
 use crate::core::{self, RequestOptions};
-use kanidm_unix_common::unix_passwd::{EtcGroup, EtcUser};
 use libnss::interop::Response;
+use sparkle_unix_common::unix_passwd::{EtcGroup, EtcUser};
 
 impl RequestOptions {
     fn fallback_fixture() -> Self {
