+++ unix_integration/pam_sparkle_common/src/pam/mod.rs
@@ -31,15 +31,15 @@ pub mod items;
 pub mod macros;
 pub mod module;
 
-use std::collections::BTreeSet;
-use std::convert::TryFrom;
-use std::ffi::CStr;
-use sparkle_unix_common::constants::DEFAULT_CONFIG_PATH;
-use sparkle_unix_common::unix_config::PamNssConfig;
 use crate::core::{self, RequestOptions};
 use crate::pam::constants::*;
 use crate::pam::module::{PamHandle, PamHooks};
 use constants::PamResultCode;
+use sparkle_unix_common::constants::DEFAULT_CONFIG_PATH;
+use sparkle_unix_common::unix_config::PamNssConfig;
+use std::collections::BTreeSet;
+use std::convert::TryFrom;
+use std::ffi::CStr;
 use time::OffsetDateTime;
 use tracing::debug;
 use tracing_subscriber::filter::LevelFilter;
