+++ unix_integration/pam_sparkle_common/src/tests.rs
@@ -3,8 +3,8 @@ use crate::core::PamHandler;
 use crate::core::{self, RequestOptions};
 use crate::module::PamResult;
 use crate::pam::ModuleOptions;
-use kanidm_unix_common::unix_passwd::{CryptPw, EtcShadow, EtcUser};
-use kanidm_unix_common::unix_proto::{DeviceAuthorizationResponse, PamServiceInfo};
+use sparkle_unix_common::unix_passwd::{CryptPw, EtcShadow, EtcUser};
+use sparkle_unix_common::unix_proto::{DeviceAuthorizationResponse, PamServiceInfo};
 use std::collections::VecDeque;
 use std::str::FromStr;
 use std::sync::Mutex;
