+++ unix_integration/resolver_common/Cargo.toml
@@ -16,6 +16,8 @@ default = ["unix"]
 unix = []
 selinux = ["sparkle_unix_common/selinux"]
 tpm = ["kanidm-hsm-crypto/tpm", "sparkle_unix_common/tpm"]
+dhat-heap = ["dep:dhat"]
+dhat-ad-hoc = ["dep:dhat"]
 
 [lib]
 name = "sparkle_resolver_common"
@@ -78,37 +80,6 @@ tempfile = { workspace = true }
 clap = { workspace = true, features = ["derive"] }
 clap_complete = { workspace = true }
 kanidm_build_profiles = { workspace = true }
-
-## Debian packaging
-# We only package unixd itself here as shared libraries must be packaged separarately due to multiarch rules.
-[package.metadata.deb]
-name = "kanidm-unixd"
-maintainer = "James Hodgkinson <james@terminaloutcomes.com>"
-depends = [
-	"libc6",
-	"tpm-udev",
-	"libpam-kanidm",
-	"libnss-kanidm"
-]
-section = "network"
-priority = "optional"
-changelog = "../../target/debian/changelog"  # Generated by platform/debian/build_debs.sh
-assets = [
-    [ "target/release/kanidm_ssh_authorizedkeys", "usr/sbin/", "755" ],
-    [ "target/release/kanidm_unixd_tasks", "usr/sbin/", "755" ],
-    [ "target/release/kanidm_unixd", "usr/sbin/", "755" ],
-    [ "target/release/kanidm-unix", "usr/sbin/", "755" ],
-    [ "../../examples/unixd-safe-default", "etc/kanidm/unixd", "644" ],
-    [ "../../examples/kanidm-safe-default", "etc/kanidm/config", "644" ],
-    [ "../../examples/kanidm", "usr/share/kanidm-unixd/", "444" ],
-    [ "../../examples/unixd", "usr/share/kanidm-unixd/", "444" ],
-    [ "../../examples/sshd-config.conf", "etc/ssh/sshd_config.d/10-kanidm.conf", "644"],
-]
-maintainer-scripts = "debian/"
-systemd-units = [
-	{ unit-name = "kanidm-unixd", enable = true},
-	{ unit-name = "kanidm-unixd-tasks", enable = true},
-]
 
 [package.metadata.cargo-machete]
 ignored = ["kanidm_build_profiles", "clap_complete"]
