+++ unix_integration/resolver_common/build.rs
@@ -1,4 +1,3 @@
-
 fn main() {
     profiles::apply_profile();
 }
