+++ unix_integration/resolver_common/src/cli/tool.rs
@@ -1,15 +1,15 @@
-use std::process::ExitCode;
-use clap::Parser;
+use crate::check::check_nsswitch_has_module;
+use crate::opt::tool::{KanidmUnixOpt, KanidmUnixParser};
 use crate::SparkleFlavour;
+use clap::Parser;
 use sparkle_unix_common::client::DaemonClient;
 use sparkle_unix_common::constants::DEFAULT_CONFIG_PATH;
 use sparkle_unix_common::unix_config::PamNssConfig;
 use sparkle_unix_common::unix_proto::{
     ClientRequest, ClientResponse, PamAuthRequest, PamAuthResponse, PamServiceInfo,
 };
-use crate::check::check_nsswitch_has_module;
 use std::path::PathBuf;
-use crate::opt::tool::{KanidmUnixParser, KanidmUnixOpt};
+use std::process::ExitCode;
 
 macro_rules! setup_client {
     () => {{
@@ -266,4 +266,3 @@ pub async fn main<F: SparkleFlavour>(flavour: F) -> ExitCode {
         }
     }
 }
-
