+++ unix_integration/resolver_common/src/idprovider/system.rs
@@ -1,10 +1,10 @@
-use super::interface::{AuthCredHandler, AuthRequest, Id, IdpError};
-use hashbrown::HashMap;
 use super::interface::{
     unix_common::unix_passwd::{CryptPw, EtcGroup, EtcShadow, EtcUser},
     unix_common::unix_proto::PamAuthRequest,
     unix_common::unix_proto::{NssGroup, NssUser},
 };
+use super::interface::{AuthCredHandler, AuthRequest, Id, IdpError};
+use hashbrown::HashMap;
 use std::sync::Arc;
 use time::OffsetDateTime;
 use tokio::sync::Mutex;
