+++ unix_integration/resolver_common/src/resolver.rs
@@ -19,6 +19,7 @@ use crate::idprovider::system::{
 use hashbrown::HashMap;
 use kanidm_hsm_crypto::provider::BoxedDynTpm;
 use kanidm_lib_file_permissions::diagnose_path;
+use lru::LruCache;
 use sparkle_unix_common::constants::{
     DEFAULT_CACHE_TIMEOUT_JITTER_MS, DEFAULT_CACHE_TIMEOUT_MAXIMUM, DEFAULT_CACHE_TIMEOUT_MINIMUM,
     DEFAULT_SHELL_SEARCH_PATHS, SYSTEM_SHADOW_PATH,
@@ -29,7 +30,6 @@ use sparkle_unix_common::unix_proto::{
     HomeDirectoryInfo, NssGroup, NssUser, PamAuthRequest, PamAuthResponse, PamServiceInfo,
     ProviderStatus,
 };
-use lru::LruCache;
 use std::fmt::Display;
 use std::num::NonZeroUsize;
 use std::ops::DerefMut;
